Option for Date(Time) objects be **in a timezone**
Often times a layperson using a dashboard can be misled when they look
at a record's creation/update time and not realize it's rooted in the
UTC timezone. This commit adds the option to display `Field::DateTime`
fields as being in a provided timezone.

Also - update the documentation to contain the two options for that
field: `:format` and `:timezone`.

Note: By default the timezone in that field is explicitly set as UTC.